requirements.txt
chromadb
fastapi
google-generativeai
joblib
kagglehub
matplotlib
numpy
pandas
pydantic
pypdf
python-dotenv
python-multipart
scikit-learn
scipy
sdv
seaborn
shap
uvicorn
